:recycle: refactor(users): split permissions server and ui code
