+++ app/core/utils/index.ts
@@ -6,3 +6,4 @@ export {
 export { NonceContext, NonceProvider, useNonce } from './nonce/nonce.ts';
 export { type AuthProvider } from './provider/provider.ts';
 export * from './monitoring/monitoring.client.ts';
+export * from './user/user.ts';
